api/requirements.txt
fastapi==0.110.0
uvicorn[standard]
pydantic-settings==2.2.1
pydantic[email]
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
httpx==0.27.0
loguru==0.7.2
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4

# DEV
isort==5.13.2
